test: Add tests for setEModeCategory

+++ contracts/mocks/helpers/MockReserveConfiguration.sol
@@ -103,12 +103,12 @@ contract MockReserveConfiguration {
     return configuration.getEModeCategory();
   }
 
-  function setEModeCategory(uint8 categoryId) external {
+  function setEModeCategory(uint256 categoryId) external {
     DataTypes.ReserveConfigurationMap memory config = configuration;
     config.setEModeCategory(categoryId);
     configuration = config;
   }
-  
+
   function setSupplyCap(uint256 supplyCap) external {
     DataTypes.ReserveConfigurationMap memory config = configuration;
     config.setSupplyCap(supplyCap);
